supported both JS and rsgan loss in RebarGAN

+++ utils/data_loader.py
@@ -127,7 +127,7 @@ class DisDataIter:
     def prepare(self, pos_samples, neg_samples, gpu=False):
         """Build inp and target"""
         inp = torch.cat((pos_samples, neg_samples), dim=0).long().detach()  # !!!need .detach()
-        target = torch.ones(inp.size(0)).long()
+        target = torch.ones(inp.size(0)).float()
         target[pos_samples.size(0):] = 0
 
         # shuffle
